Made changes to pagination
Made changes to pagination and corrected errors
