+++ src/redux/auth/authOperations.js
@@ -4,8 +4,8 @@ import { toast } from 'react-toastify';
 import { IoMdLogIn } from 'react-icons/io';
 import { store } from 'redux/store';
 
-// axios.defaults.baseURL = 'https://petly-back.onrender.com/api';
-axios.defaults.baseURL = 'http://localhost:3001/api/';
+axios.defaults.baseURL = 'https://petly-back.onrender.com/api';
+// axios.defaults.baseURL = 'http://localhost:3001/api/';
 
 const token = {
   set(token) {
